ProactiveRefresh -> GetTokenMixin
